🔨 Refactor: 로직 일부 수정
